backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "somameet-backend"
version = "0.1.0"
description = "SomaMeet MVP backend - privacy-preserving meeting scheduler"
requires-python = ">=3.11"
authors = [
    {name = "SomaMeet Team"},
]
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.7",
    "icalendar>=5.0",
    "httpx>=0.26",
    "google-auth>=2.27",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.116",
    "google-generativeai>=0.4",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.18"]
openai = ["openai>=1.10"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
    "freezegun>=1.4",
    "ruff>=0.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
    "acceptance: end-to-end acceptance tests",
    "integration: integration tests",
    "unit: unit tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
